Refactor rrevent constructor to use size_t for loop index; update rrblemappings to use Arduino pin constants; modify move_t function signature and implementation; adjust main loop buffer termination; enhance serialization in serde.

+++ lib/rrevent/include/rrevent.hpp
@@ -16,7 +16,7 @@ class rrevent {
      * @param mcount The number of elements in the data array to copy.
      */
     rrevent(const int cmd, const size_t sz, const String data[]) : _cmd(cmd), _sz(sz) {
-        for (int i = 0; i < sz; i++) {
+        for (size_t i = 0; i < sz; i++) {
             _data[i] = String(data[i].c_str());
         }
     }
